gameloop improved smooth ball added zoom fixed positial correction removed verlet integration (test) replaced Stack by ArrayDeque added other manifold update methods determinant for 3x3 matrix added

+++ src/com/brocorporation/gameengine/IUpdateInfo.java
@@ -2,7 +2,7 @@ package com.brocorporation.gameengine;
 
 public interface IUpdateInfo {
 
-	public float getInverseRate();
+	public int getInverseRate();
 	
 	public float getRate();
